wms/requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0
